开始调试登录接口

+++ app/admin/src/ctls/mod.rs
@@ -5,12 +5,12 @@ pub mod user;
 
 #[derive(Debug, serde::Serialize, serde::Deserialize)]
 pub struct Claims {
-    user_id: i64,
+    user_id: i32,
     exp: i128,
 }
 
 impl Claims {
-    pub fn build(user_id: i64) -> Self {
+    pub fn build(user_id: i32) -> Self {
         Self {
             user_id,
             exp: time::OffsetDateTime::now_utc().unix_timestamp_nanos(),
